docs(PPDSC-2420): introduce new homepage (#376)
* docs(PPDSC-2420): introduce new homepage
