--- site/components/illustrations/landing-page/hero-image.tsx
@@ -1,87 +1,0 @@
-import React from 'react';
-import {styled} from 'newskit';
-import {Path} from '../path';
-
-const Svg = styled.svg`
-  .dot1 {
-    transform: translate(140px, -80px);
-  }
-  .dot2 {
-    transform: translate(0px, 95px);
-  }
-  .dot3 {
-    transform: translate(-25px, -45px);
-  }
-`;
-
-export const HeroImage: React.FC = () => (
-  <Svg viewBox="155 135 850 920" fill="none" xmlns="http://www.w3.org/2000/svg">
-    <Path
-      fillRule="evenodd"
-      clipRule="evenodd"
-      d="M909.317 855.735C1007.5 714 1007.5 624.5 1000.5 540.5C993.5 456.5 969.5 402.5 903 352.5C836.5 302.5 870.5 241.5 776.999 173.5C683.499 105.5 425.5 153 333.501 234.5C241.501 316 323.588 399.771 263.286 500.019C201.614 602.545 137.347 679.122 167.159 782.22C196.97 885.317 353.112 969.519 457.064 1027.13C558.424 1083.29 762.068 1055.84 909.317 855.735Z"
-      fill="illustrationBackground020"
-    />
-    <Path
-      d="M448.423 467.131C451.703 467.131 454.362 464.581 454.362 461.435C454.362 458.289 451.703 455.738 448.423 455.738C445.143 455.738 442.484 458.289 442.484 461.435C442.484 464.581 445.143 467.131 448.423 467.131Z"
-      fill="illustrationInterface020"
-    />
-    <Path
-      d="M312.131 960.922C316.784 969.683 327.96 973.166 337.092 968.702C346.225 964.239 349.856 953.519 345.203 944.758C340.55 935.998 329.374 932.515 320.241 936.979C311.109 941.442 307.478 952.162 312.131 960.922Z"
-      fill="illustrationInterface020"
-    />
-    <Path
-      className="dot3"
-      d="M852.543 1098.17C871.864 1098.17 887.527 1083.15 887.527 1064.61C887.527 1046.08 871.864 1031.05 852.543 1031.05C833.222 1031.05 817.559 1046.08 817.559 1064.61C817.559 1083.15 833.222 1098.17 852.543 1098.17Z"
-      fill="illustrationInterface020"
-    />
-    <Path
-      className="dot1"
-      d="M298.113 444.627C316.562 444.627 331.519 430.281 331.519 412.583C331.519 394.886 316.562 380.54 298.113 380.54C279.663 380.54 264.707 394.886 264.707 412.583C264.707 430.281 279.663 444.627 298.113 444.627Z"
-      fill="illustrationInterface020"
-    />
-    <Path
-      className="dot2"
-      d="M321.367 520.095C331.617 520.095 339.926 512.125 339.926 502.293C339.926 492.461 331.617 484.491 321.367 484.491C311.118 484.491 302.809 492.461 302.809 502.293C302.809 512.125 311.118 520.095 321.367 520.095Z"
-      fill="illustrationInterface020"
-    />
-    <Path
-      fillRule="evenodd"
-      clipRule="evenodd"
-      d="M403.395 616.805C361.454 726.062 405.318 781.039 406.238 782.457C433.103 823.863 465.622 853.402 537.824 865.929C610.025 878.455 648.128 857.453 694.237 832.422C740.346 807.39 730.686 757.005 745.516 703.859C760.684 649.506 798.731 598.161 771.834 554.345C744.937 510.529 644.111 498.901 589.012 484.724C535.29 470.902 445.336 507.547 403.395 616.805Z"
-      fill="illustrationInterface090"
-    />
-    <Path
-      fillRule="evenodd"
-      clipRule="evenodd"
-      d="M358.227 865.503C435.898 954.973 507.465 946.023 509.204 945.968C560.02 944.353 602.949 932.11 650.359 878.393C697.769 824.677 697.859 782.524 698.314 731.704C698.769 680.885 648.45 663.717 607.883 624.824C566.395 585.048 539.063 527.769 486.056 528.205C433.048 528.641 372.137 606.584 331.789 645.268C292.448 682.984 280.556 776.034 358.227 865.503Z"
-      fill="blue020"
-    />
-    <Path
-      fillRule="evenodd"
-      clipRule="evenodd"
-      d="M658.327 830.989C684.94 761.667 657.099 726.783 656.515 725.883C639.463 699.61 616.28 663.164 570.456 655.213C524.632 647.262 463.758 682.32 462.508 683.157C437.783 699.715 450.152 743.64 441.195 775.738C431.572 810.225 424.692 839.613 441.764 867.415C458.837 895.218 505.562 905.788 540.531 914.785C574.627 923.558 631.715 900.31 658.327 830.989Z"
-      fill="blue030"
-    />
-    <Path
-      d="M302.436 669.316C278.874 685.488 265.34 704.384 265.34 724.572C265.34 784.306 383.827 832.731 529.989 832.731C676.151 832.731 794.638 784.306 794.638 724.572C788.278 705.562 772.697 696.057 747.894 696.057"
-      stroke="illustrationInterface100"
-      strokeWidth="4"
-      strokeDasharray="8 20"
-    />
-    <Path
-      d="M381.747 841.356C364.938 841.356 351.311 828.285 351.311 812.161C351.311 796.037 364.938 782.965 381.747 782.965C398.557 782.965 412.184 796.037 412.184 812.161C412.184 828.285 398.557 841.356 381.747 841.356Z"
-      fill="teal040"
-    />
-    <Path
-      d="M377.472 890.684C384.28 917.53 397.515 936.938 417.284 945.781C481.654 974.573 590.934 880.459 661.368 735.571C731.802 590.683 736.718 449.887 672.348 421.095C633.184 403.577 577.395 431.556 523.845 488.787"
-      stroke="illustrationInterface100"
-      strokeWidth="4"
-      strokeDasharray="11 20"
-    />
-    <Path
-      d="M710.691 489.908C699.826 489.908 691.019 481.459 691.019 471.037C691.019 460.616 699.826 452.167 710.691 452.167C721.556 452.167 730.363 460.616 730.363 471.037C730.363 481.459 721.556 489.908 710.691 489.908Z"
-      fill="blue040"
-    />
-  </Svg>
-);
